templates/pipeline/requirements.txt
# Pipeline template dependencies
sentence-transformers>=2.2
lancedb>=0.4
bertopic>=0.15
hdbscan>=0.8
umap-learn>=0.5
pandas>=2.0
numpy>=1.24
scipy>=1.10
scikit-learn>=1.3
torch>=2.0
transformers>=4.30
networkx>=3.0
python-louvain>=0.16
plotly>=5.15
streamlit>=1.30
pyvis>=0.3
anthropic>=0.20
